Conflicts:

	common/network/network_model.cc
	common/network/network_model_emesh_hop_by_hop.cc
	common/performance_model/dram_perf_model.cc
	common/performance_model/queue_model_history_list.cc
	common/performance_model/queue_model_history_list.h

+++ tests/unit/shared_mem_test4/shared_mem_test4.cc
@@ -61,7 +61,7 @@ int main (int argc, char *argv[])
 
 void* thread_func(void* threadid)
 {
-   int tid = (int) threadid;
+   long tid = (long) threadid;
    Core* core = Sim()->getCoreManager()->getCurrentCore();
 
    for (int i = 0; i < num_iterations; i++)
